* "Make `ContextImpl` destructor `= default` to simplify implementation…

* "Update references to TLS-specific security configuration, replace `t…

* "Mark `writePKCS12File` as `PVXS_API` for consistent symbol export an…

* "Remove GCC diagnostic pragmas around tautological comparison in `ser…

* "Replace Windows 2019 builds with Windows 2022 using VS2022 in CI wor…

* "Improve error handling in `certcontext.h` by checking `getcwd()` ret…

* "Standardize formatting and naming conventions in SPVA authenticator …

* "True isolation for concurrently run tests involving Mock CMS."

* "Try to remediate SEGV by switching order of SSL_CTX manipulation."

* "cherry pick edcc21bca35502406b5b80e5d253c67b9a195bf9"

* "fix ci to be in line with master wit exception of openssl changes"

* "Turn on DEBUG and RUN!!!!!!!!"

* "Turn off DEBUG and reorganise events and guards for handling TLS con…

* "Sort out globals lurking in WildcardPV"

* fix ci error for make status

* Add back nofortify logic to ci

* Add extra CPPFLAGS to CI build configuration

* Comment out OpenSSL build flag in CI workflows

* "Refactor certificate status manager to use passed in client config f…

* "Replace `std::cout` usage with logger utilities, clean up unused inc…

* Update CI workflows to adjust build configurations and remove redunda…

* Update CI workflows to adjust build configurations and remove redunda…

* Remove conditional OpenSSL handling for isTLS in connection classes

* Remove conditional OpenSSL handling for isTLS in `Connection` constru…

* Remove redundant OpenSSL preprocessor directives and simplify related…

* "Clean up unused includes in `config.cpp` to improve maintainability …

* "Remove unused `<pwd.h>` include from `util.cpp` to clean up dependen…

* "Replace `#ifdef __unix__` with `#ifndef _WIN32` in `util.cpp` to imp…

* "Reorder includes in `util.cpp` to ensure `libgen.h` is only included…

* "Reorder includes in `util.cpp` to ensure platform-specific headers (…

* "Add `_WIN32` guard for `PATH_MAX` definition and adjust preprocessor…

* Replace `#ifdef __unix__` with `#ifndef _WIN32` in `util.cpp` for con…

* libevent: override minimum cmake version for bundled build

* Include `<direct.h>` for `_WIN32` instead of defining unused `PATH_MA…

* Refactor `XDG` helper functions to remove default argument usage and …

* Simplify `simpleget.cpp` example by removing threading logic, redunda…

* missing include

* Merge remote-tracking branch 'origin/tls-dev' into tls-dev

* revert bundled libevent

* Try enabling `PVXS_ENABLE_PVACMS` in Native Linux 7.0-secure-pvaccess…

* Fix PVACMS CI build errors

* Fix PVACMS CI build errors and add Authenticators

* remove Authenticators from native builds due to missing dependencies

* Extend mock source to track peer requests and update related counters…

* Extend mock source to track peer requests and update related counters…

* Add thread safety to cert status counters and streamline counter mana…

* Add runtime OpenSSL and libevent link test to configure step

* Change certificate status message and CCR response message to be NTEn…

* Refactor cert status request counters to simplify structure and impro…

* reset the openssl probe to original

* Change signature of subscribeToPeerStatus to be noexcept to allow pro…

* Change cmake flags for cygwin builds to search for libraries with the…

* Update CMake library suffix for cygwin builds to use `.a` instead of …

* Update CMake flags for Windows builds to use `.lib` suffix for static…

* Simplify CMake flags for Windows builds by removing unnecessary type …

+++ certs/pvacms.h
@@ -278,7 +278,7 @@ class StatusMonitor {
         std::vector<serial_number_t> result;
         Guard G(lock_);
         for (const auto &pair : active_status_validity_) {
-            if (pair.second > cutoff) {
+            if (static_cast<uint64_t>(pair.second) > cutoff) {
                 result.push_back(pair.first);
             }
         }
